update expected goal model application for easier use

+++ src/database/sql_queries/create_events_db.sql
@@ -6,15 +6,15 @@ CREATE TABLE public.events
 (
     game_id varchar(80),
     event_id varchar(80),
-    minute int,
-    second int,
+    minute float,
+    second float,
     team_id varchar(80),
     player_id varchar(80),
-    x int,
-    y int,
-    type_value int,
+    x float,
+    y float,
+    type_value float,
     type_name varchar(80),
-    outcome_type_value int,
+    outcome_type_value float,
     outcome_type_name varchar(80),
-    is_touch int
+    is_touch float
 );
